backend_flask/requirements.txt
wheel
argon2-cffi
flask
flask-jwt-extended
marshmallow-sqlalchemy
python-dotenv
sqlalchemy
flask-socketio
mysqlclient
openpyxl
gunicorn